Add an algorithm for fibonacci partial sum test

+++ algorithmic_toolbox/w2/6/fibonacci_sum_last_digit_test.cpp
@@ -5,28 +5,28 @@
 using std::vector;
 using std::cout;
 
-long long fibonacci_sum_fast(long long n) {
-    long long num0 = 0;
-    long long num1 = 1;
+int fibonacci_sum_fast(long long n) {
+    int num0 = 0;
+    int num1 = 1;
 
     if (n == 0)
         return 0;
     else if (n == 1)
         return 1;
     else {
-        long long remainder = n % 60;
+        int remainder = n % 60;
 
         if (remainder == 0)
-          return 0;
+            return 0;
 
-        for (long long i = 2; i < remainder + 3; ++i) {
-          long long digit = (num0 + num1) % 60;
-          num0 = num1;
-          num1 = digit;
+        for (int i = 2; i < remainder + 3; ++i) {
+            int digit = (num0 + num1) % 60;
+            num0 = num1;
+            num1 = digit;
         }
     }
 
-    long long sum = num1 - 1;
+    int sum = num1 - 1;
 
     return sum % 10;
 }
